Update README.md
This will limit the action to to only work on js type files. 

I found that if you don't limit the binding to js files it will mess up your HTML. It's also useful for allowing you to bind another formatter to the same key-binding for other file types (IE HTML).